剩下最后7个错误

+++ programs/bumpin-trade/src/instructions/pool/pool_stake.rs
@@ -81,7 +81,7 @@ pub struct StakeParams {
     portfolio: bool,
 }
 
-pub fn handle_pool_stake(mut ctx: Context<PoolStake>, pool_index: u16, trade_token_index: u16, stake_params: StakeParams) -> anchor_lang::Result<()> {
+pub fn handle_pool_stake(mut ctx: Context<PoolStake>, pool_index: usize, trade_token_index: u16, stake_params: StakeParams) -> anchor_lang::Result<()> {
     let mut pool = &mut ctx.accounts.pool.load_mut()?;
     validate!(pool.pool_config.mini_stake_amount>stake_params.request_token_amount, BumpErrorCode::StakeToSmall);
